% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareNA.R
\name{compareNA}
\alias{compareNA}
\title{Compare a value to NA}
\usage{
compareNA(v1, v2)
}
\arguments{
\item{v1}{Variable 1}

\item{v2}{Variable 2}
}
\value{
TRUE if v1 or v2 are both non-NA or NA. FALSE if they do ot match.
}
\description{
Compare a value to NA
}
