% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSexAgeCateg.R
\name{createSexAgeCateg}
\alias{createSexAgeCateg}
\title{Create a vector of age and sex categories}
\usage{
createSexAgeCateg(ageVec, sexVec, categoryNum)
}
\arguments{
\item{ageVec}{Vector containing the age of participants.
Make sure the ID order is identical to sexVec.}

\item{sexVec}{Vector containing the gender of participants.
Make sure the ID order is identical to ageVec.}

\item{categoryNum}{Number of categories of each gender to create.
Currently supports options for categories of
size 8 or 16.}
}
\value{
Vector of age and sex category by ID
}
\description{
Create a vector of age and sex categories
}
