% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridArrangeSharedLegend.R
\name{gridArrangeSharedLegend}
\alias{gridArrangeSharedLegend}
\title{Grid Arrange, Shared Legend}
\usage{
gridArrangeSharedLegend(..., ncol = length(list(...)), nrow = 1,
  position = c("bottom", "right"))
}
\arguments{
\item{...}{Plots to be arranged in grid}

\item{ncol}{Number of columns in grid}

\item{nrow}{Number of rows in grid}

\item{position}{Gird position (bottom or right)}
}
\value{
Grid of plots with common legend
}
\description{
Plot Multiple Graphs with Shared Legend in a Grid
}
\author{
Phil Boileau, \email{philippe.boileau (at) rimuhc.ca}
}
