% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recodeChronicConditions.R
\name{recodeChronicConditions}
\alias{recodeChronicConditions}
\title{Recode Chronic Conditions}
\usage{
recodeChronicConditions(variable)
}
\arguments{
\item{variable}{CCC or CCT variable that will be recoded}
}
\value{
Recoded variable
}
\description{
Note the mapping function: 1 = "Yes", 2 = "No",
(8, 9, -8) = "DK/NoAn/REF/MIS"
}
