% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryTable.R
\name{summaryTable}
\alias{summaryTable}
\title{Create stratified summary tables}
\usage{
summaryTable(vect, categories, catname, missingValues = c())
}
\arguments{
\item{vect}{The vector of data to be summarized}

\item{categories}{The vector of categories}

\item{catname}{The name of the categories}

\item{missingValues}{Optionally, the vector containing the codes of the missing values.}
}
\value{
Matrix of stratified summaries
}
\description{
Create stratified summary tables
}
\examples{
\dontrun{
summaryTable(tra3.2$COG_AFT_SCORE_2_TRM, tra3.2$SEX_ASK_TRM, "Gender")
# produces a summary table of COG_AFT_SCORE_2_TRM stratified by gender
}
}
