% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgtSummaryTable.R
\name{wgtSummaryTable}
\alias{wgtSummaryTable}
\title{Create weighted, stratified summary tables}
\usage{
wgtSummaryTable(vect, wgts, categories, catname, missingValues = c())
}
\arguments{
\item{vect}{The vector of data to be summarized}

\item{wgts}{The vector containing the weights}

\item{categories}{The vector of categories}

\item{catname}{The name of the categories}

\item{missingValues}{Optionally, the vector containing the codes of the missing values}
}
\value{
Matrix of stratified summaries
}
\description{
Create weighted, stratified summary tables
}
