% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadCLSAData.R
\name{loadCLSAData}
\alias{loadCLSAData}
\title{Load CLSA Comprehensive and Tracking Datasets}
\usage{
loadCLSAData(path)
}
\arguments{
\item{path}{The path to the folder containging the comprehensive and
tracking .csv files}
}
\value{
A list of the comprehensive and tracking CLSA datasets in the following order:
\enumerate{
\item comprehensive
\item tracking
}
}
\description{
Load the most recent comprehensive and tracking data sets. Note that
the .csv files must already be saved on your computer before using this
function. Make sure that the csv files containing the data are named
as follows:
\itemize{
\item comprehensive: cop3_2.csv
\item tracking: tra3_3.csv
}
}
\examples{
 
\dontrun{
dataList <- loadCLSAData(path = "C:/Users/Documents")

# tracking data set
cop3.2 <- dataList[[1]]
tra3.3 <- dataList[[2]]
}
}
\author{
Phil Boileau, \email{philippe.boileau (at) rimuhc.ca}
}
