% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryTable.R
\name{summaryTable}
\alias{summaryTable}
\title{Create Stratified Summary Tables}
\usage{
summaryTable(vect, categories = c(), catname = "", missingValues = c(),
  weights = c())
}
\arguments{
\item{vect}{The vector of data to be summarized. It should be a continuous variable.}

\item{categories}{The vector of categories. If left undefined, a simple summary of
the vect vector will be performed.}

\item{catname}{The name of the categories. Defaults to the name of the category vector.}

\item{missingValues}{An optional vector containing the codes of the missing values.}

\item{weights}{An optional vector containing the weights of the data.}
}
\value{
Returns a table of stratified summaries.
}
\description{
Creates a summary table stratified by a categorical variable. If being used for the generation of
a report, it is recommended to use the \code{\link[pander]{pander}} function to format the table.
Supports the use of weighted data.
}
\author{
Phil Boileau, \email{philippe.boileau (at) rimuhc.ca}
}
